@@ -6,12 +6,11 @@ description: Demonstrates how to use the float_button_overlay plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  package_info: ^0.4.3+2
-  path_provider: ^0.5.0+1
-  dio: ^3.0.10  
+  package_info: ^2.0.2
+  path_provider: ^2.0.3
   flutter:
     sdk: flutter
 
