@@ -4,8 +4,7 @@ version: 0.3.4
 homepage: https://github.com/niltondiniz/float_button_overlay
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0"
+  sdk: ">=2.12.0 <3.0.0"  
 
 dependencies:
   flutter:
